+++ mybatis-plus-extension/src/main/java/com/baomidou/mybatisplus/extension/spi/CompatibleHelper.java
@@ -15,20 +15,69 @@
  */
 package com.baomidou.mybatisplus.extension.spi;
 
+import com.baomidou.mybatisplus.core.toolkit.Assert;
+import org.apache.ibatis.logging.Log;
+import org.apache.ibatis.logging.LogFactory;
+
 import java.util.ServiceLoader;
 
 /**
  * 兼容处理辅助类
+ * <p>默认加载使用SPI实现,需要手动指定请使用{@link #setCompatibleSet(CompatibleSet)}</p>
+ *
+ * @see com.baomidou.mybatisplus.core.spi.CompatibleHelper
+ * @deprecated 3.5.12
  */
+@Deprecated
 public class CompatibleHelper {
 
-    private static CompatibleSet COMPATIBLE_SET;
+    private static final Log LOG = LogFactory.getLog(CompatibleHelper.class);
 
-    public static CompatibleSet getCompatibleSet() {
-        if (null == COMPATIBLE_SET) {
-            ServiceLoader<CompatibleSet> loader = ServiceLoader.load(CompatibleSet.class);
-            COMPATIBLE_SET = loader.iterator().next();
+    private static CompatibleSet COMPATIBLE_SET = null;
+
+    static {
+        ServiceLoader<CompatibleSet> loader = ServiceLoader.load(CompatibleSet.class, CompatibleSet.class.getClassLoader());
+        int size = 0;
+        for (CompatibleSet compatibleSet : loader) {
+            size++;
+            LOG.debug("Load compatibleSet: " + compatibleSet);
+            COMPATIBLE_SET = compatibleSet;
         }
+        if (size > 1) {
+            LOG.warn("There are currently multiple implementations, and the last one is used " + COMPATIBLE_SET);
+        }
+    }
+
+    /**
+     * 判断是否存在 {@link com.baomidou.mybatisplus.extension.spi.CompatibleSet} 实例
+     *
+     * @return 是否存在 (存在返回true,为空返回false)
+     * @since 3.5.12
+     */
+    public static boolean hasCompatibleSet() {
+        return COMPATIBLE_SET != null;
+    }
+
+    /**
+     * 手动指定 {@link com.baomidou.mybatisplus.extension.spi.CompatibleSet} 实例
+     *
+     * @param compatibleSet {@link com.baomidou.mybatisplus.extension.spi.CompatibleSet} 实例
+     * @since 3.5.12
+     */
+    public static void setCompatibleSet(CompatibleSet compatibleSet) {
+        COMPATIBLE_SET = compatibleSet;
+    }
+
+    /**
+     * 获取{@link com.baomidou.mybatisplus.extension.spi.CompatibleSet}实例
+     * <p>当为空时会抛出异常,需要检查是否为空请使用{@link #hasCompatibleSet()}</p>
+     *
+     * @return {@link com.baomidou.mybatisplus.extension.spi.CompatibleSet}
+     * @see #setCompatibleSet(CompatibleSet)
+     */
+    public static CompatibleSet getCompatibleSet() {
+        Assert.isTrue(hasCompatibleSet(), "Please add specific implementation dependencies or use the setCompatibleSet method to specify");
         return COMPATIBLE_SET;
     }
+
 }