+++ mybatis-plus-jsqlparser-support/mybatis-plus-jsqlparser/src/main/java/com/baomidou/mybatisplus/extension/plugins/inner/BaseMultiTableInnerInterceptor.java
@@ -17,6 +17,7 @@ package com.baomidou.mybatisplus.extension.plugins.inner;
 
 import com.baomidou.mybatisplus.core.toolkit.CollectionUtils;
 import com.baomidou.mybatisplus.extension.parser.JsqlParserSupport;
+import com.baomidou.mybatisplus.jsqlparser.enums.ExpressionAppendMode;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
@@ -48,6 +49,13 @@ import java.util.stream.Collectors;
 @SuppressWarnings({"rawtypes"})
 public abstract class BaseMultiTableInnerInterceptor extends JsqlParserSupport implements InnerInterceptor {
 
+    /**
+     * 条件表达式追加模式 (默认放置最后,仅作用于update,delete,select)
+     *
+     * @since 3.5.11
+     */
+    private ExpressionAppendMode expressionAppendMode = ExpressionAppendMode.LAST;
+
     protected void processSelectBody(Select selectBody, final String whereSegment) {
         if (selectBody == null) {
             return;
@@ -77,9 +85,9 @@ public abstract class BaseMultiTableInnerInterceptor extends JsqlParserSupport i
         }
         if (where != null) {
             if (where instanceof OrExpression) {
-                return new AndExpression(new ParenthesedExpressionList<>(where), expression);
+                return appendExpression(new ParenthesedExpressionList<>(where), expression);
             } else {
-                return new AndExpression(where, expression);
+                return appendExpression(where, expression);
             }
         }
         return expression;
@@ -181,6 +189,9 @@ public abstract class BaseMultiTableInnerInterceptor extends JsqlParserSupport i
                 Expression inExpression = expression.getRightExpression();
                 if (inExpression instanceof Select) {
                     processSelectBody(((Select) inExpression), whereSegment);
+                } else if (inExpression instanceof AndExpression) {
+                    Expression leftExpression = ((AndExpression) inExpression).getLeftExpression();
+                    processWhereSubSelect(leftExpression, whereSegment);
                 }
             } else if (where instanceof ExistsExpression) {
                 // exists
@@ -403,9 +414,25 @@ public abstract class BaseMultiTableInnerInterceptor extends JsqlParserSupport i
             return injectExpression;
         }
         if (currentExpression instanceof OrExpression) {
-            return new AndExpression(new ParenthesedExpressionList<>(currentExpression), injectExpression);
+            return appendExpression(new ParenthesedExpressionList<>(currentExpression), injectExpression);
         } else {
+            return appendExpression(currentExpression, injectExpression);
+        }
+    }
+
+    /**
+     * 追加表达式，默认追加到后面，可以配置变量 {@link #expressionAppendMode} 来控制追加到前面还是后面
+     *
+     * @param currentExpression 原sql的条件表达式
+     * @param injectExpression  注入的表达式
+     * @return 追加了条件的完整表达式(where条件 / on条件)
+     * @since 3.5.11
+     */
+    protected Expression appendExpression(Expression currentExpression, Expression injectExpression) {
+        if (ExpressionAppendMode.LAST == expressionAppendMode || expressionAppendMode == null) {
             return new AndExpression(currentExpression, injectExpression);
+        } else {
+            return new AndExpression(injectExpression, currentExpression);
         }
     }