+++ mybatis-plus-spring/src/main/java/com/baomidou/mybatisplus/extension/spi/SpringCompatibleSet.java
@@ -15,10 +15,12 @@
  */
 package com.baomidou.mybatisplus.extension.spi;
 
+import com.baomidou.mybatisplus.core.toolkit.AopUtils;
 import com.baomidou.mybatisplus.core.toolkit.ExceptionUtils;
 import lombok.SneakyThrows;
 import org.apache.ibatis.exceptions.PersistenceException;
 import org.apache.ibatis.logging.Log;
+import org.apache.ibatis.logging.LogFactory;
 import org.apache.ibatis.reflection.ExceptionUtil;
 import org.apache.ibatis.session.ExecutorType;
 import org.apache.ibatis.session.SqlSession;
@@ -26,6 +28,9 @@ import org.apache.ibatis.session.SqlSessionFactory;
 import org.mybatis.spring.MyBatisExceptionTranslator;
 import org.mybatis.spring.SqlSessionHolder;
 import org.mybatis.spring.SqlSessionUtils;
+import org.springframework.aop.framework.AopProxyUtils;
+import org.springframework.beans.factory.ObjectProvider;
+import org.springframework.context.ApplicationContext;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
 
@@ -36,6 +41,10 @@ import java.util.function.Consumer;
  * spring 兼容方法集接口实现类
  */
 public class SpringCompatibleSet implements CompatibleSet {
+
+    private static final Log LOG = LogFactory.getLog(SpringCompatibleSet.class);
+
+    public static volatile ApplicationContext applicationContext;
 
     @Override
     public SqlSession getSqlSession(SqlSessionFactory sessionFactory) {
@@ -88,4 +97,31 @@ public class SpringCompatibleSet implements CompatibleSet {
     public InputStream getInputStream(String path) throws Exception {
         return new ClassPathResource(path).getInputStream();
     }
+
+    @Override
+    public <T> T getBean(Class<T> clz) {
+        if (applicationContext != null) {
+            ObjectProvider<T> provider = applicationContext.getBeanProvider(clz);
+            return provider.getIfAvailable();
+        }
+        LOG.warn("The applicationContext property is empty. Please initialize it via the static field of applicationContext in SpringContextHolder or by calling the setApplicationContext method of MybatisSqlSessionFactoryBean.");
+        return null;
+    }
+
+    @Override
+    public Object getProxyTargetObject(Object mapper) {
+        Object result = mapper;
+        if (AopUtils.isLoadSpringAop()) {
+            while (org.springframework.aop.support.AopUtils.isAopProxy(result)) {
+                result = AopProxyUtils.getSingletonTarget(result);
+            }
+        }
+        return result;
+    }
+
+    @Override
+    public void setContext(Object context) {
+        applicationContext = (ApplicationContext) context;
+    }
+
 }