+++ spring-boot-starter/mybatis-plus-spring-boot-autoconfigure/src/main/java/com/baomidou/mybatisplus/autoconfigure/MybatisPlusAutoConfiguration.java
@@ -24,6 +24,7 @@ import com.baomidou.mybatisplus.core.handlers.PostInitTableInfoHandler;
 import com.baomidou.mybatisplus.core.incrementer.IKeyGenerator;
 import com.baomidou.mybatisplus.core.incrementer.IdentifierGenerator;
 import com.baomidou.mybatisplus.core.injector.ISqlInjector;
+import com.baomidou.mybatisplus.extension.spring.MybatisPlusApplicationContextAware;
 import com.baomidou.mybatisplus.extension.spring.MybatisSqlSessionFactoryBean;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.mapping.DatabaseIdProvider;
@@ -164,6 +165,7 @@ public class MybatisPlusAutoConfiguration implements InitializingBean {
         MybatisSqlSessionFactoryBean factory = new MybatisSqlSessionFactoryBean();
         factory.setDataSource(dataSource);
         factory.setVfs(SpringBootVFS.class);
+        factory.setApplicationContext(this.applicationContext);
         if (StringUtils.hasText(this.properties.getConfigLocation())) {
             factory.setConfigLocation(this.resourceLoader.getResource(this.properties.getConfigLocation()));
         }
@@ -373,6 +375,18 @@ public class MybatisPlusAutoConfiguration implements InitializingBean {
             logger.debug(
                 "Not found configuration for registering mapper bean using @MapperScan, MapperFactoryBean and MapperScannerConfigurer.");
         }
+    }
+
+    /**
+     * @deprecated 3.5.13
+     * @see MybatisSqlSessionFactoryBean#setApplicationContext(ApplicationContext)
+     * @return MybatisPlusApplicationContextAware
+     */
+    @Bean
+    @Deprecated
+    @ConditionalOnMissingBean(MybatisPlusApplicationContextAware.class)
+    public MybatisPlusApplicationContextAware mybatisPlusSpringApplicationContextAware() {
+        return new MybatisPlusApplicationContextAware();
     }
 
 }