+++ CHANGELOG.md
@@ -1,4 +1,63 @@
 ﻿# CHANGELOG
+## [v3.5.12] 2025.04.27
+- fix: 修复批量操作异步执行首次可能的出现`NoSuchElementException`错误
+- fix: 修复默认`SQL`解析线程池在`JVM`退出关闭导致的任务拒绝
+- fix: 修复`entity.java.btl`生成`toString`方法样式错误
+- fix: 修复`entity.java.ftl`模板类注释与导包缺少换行
+- opt: 重构`SqlRunner`执行`SQL`语句 (动态传参，不再根据参数值生成执行`SQL`)
+- opt: 增强`SqlRunner`执行(支持单参数使用`Map`({key}),`List`({index}),`JavaBean`({property})获取值)
+- opt: 改进`MybatisUtils`对自`SqlSessionFactory`的提取(支持自定义`SqlSessionTemplate`子类)
+- opt: 自动识别数据库支持`TDengine`数据库`websocket`连接
+- opt: 支持`Db`工具类对多数据源的支持
+- opt: 优化`MapperProxy`属性访问
+- opt: `CompatibleSet`接口增加`getBean`与`getProxyTargetObject`方法
+- opt: `CompatibleSet`与`CompatibleHelper`调整至`com.baomidou.mybatisplus.core.spi`包之中
+- opt: 支持手动指定`CompatibleSet`实现
+- opt: 代码生成器处理驱动返回索引信息`null`
+- opt: 代码生成器处理`PRIMARY_KEY_`为开头的主键索引情况
+- opt: 去除`entity.kt.btl`模板`@Override`注解
+- opt: 解决`serviceImpl.java.ej`生成格式不统一
+- opt: 去除`mapper.java.ftl`多余的换行生成
+- opt: 去除`entity.kt.vm`,`entity.kt.ej`,`entity.kt.btl`导包结束分隔符
+- opt: 去除`controller.java.ej`,`controller.java.vm`多余的换行
+- opt: 去除`entity.kt.btl`生成属性多余的空格
+- opt: 统一`entity.java.btl`,`entity.java.ej`,`entity.java.ftl`,`entity.java.vm` 生成的`toString`方法样式
+
+## [v3.5.11] 2025.03.23
+- fix: 修复代码生成器链式模型非`lombok`下生成了`@Accessors`注解
+- fix: 修复主键使用`UUID`执行批量删除错
+- fix: 修复`Kotlin`使用`select(predicate)`方法错误
+- fix: 修复`AbstractCaffeineJsqlParseCache`异步产生的错误
+- fix: 修复动态SQL解析包含SQL注释(--或#)导致的合并错误 (动态脚本语句不再处理换行,如果需要去除换行请自行处理)
+- fix: 修复`DataChangeRecorderInnerInterceptor`数据比较出现强转异常
+- fix: 修复`IllegalSQLInnerInterceptor`拦截插件获取`catalog`与`schema`错误
+- fix: 修复动态表解析`create table if not exists` 获取表名错误
+- fix: 修复动态表解析`create [type] index` 获取表名错误
+- feat: 新增`DynamicTableNameJsqlParserInnerInterceptor` 基于`JsqlParser`动态表处理
+- feat: 支持`DdlScript`自定义脚本运行器参数
+- feat: 支持`DdlHelper`自定义脚本运行器参数
+- feat: 支持`DdlApplicationRunner`参数配置(脚本错误处理,自定义`ScriptRunner`,多处理器执行异常是否中断)
+- feat: 支持`BaseMultiTableInnerInterceptor`指定追加条件模式 (默认条件追加至末尾,仅作用于`select`,`delete`,`update`)
+- feat: 支持生成器`Entity`指定`serialVersionUID`添加`@Serial`注解
+- feat: 支持生成器`Entity`注解(字段,类注解)自定义处理
+- feat: 支持生成器`Entity`导包自定义处理
+- feat: 支持`崖山`数据库
+- feat: 支持`Hive2`分页
+- feat: 升级`Gradle`至8.10
+- feat: 支持`DdlHelper`执行自定义异常处理
+- opt: 调整`DynamicTableNameInnerInterceptor`表处理逻辑并保证`hook`运行
+- opt: 调整`DdlScript`类方法实现(分离DDL版本记录,优化执行方法)
+- opt: 调整`DbType#GAUSS`数据库名为`gauss`
+- opt: 调整`JsqlParserGlobal`解析线程池指定
+- opt: 移除过时的`FieldStrategy.IGNORED`
+- opt: 移除过时的`GlobalConfig.DbConfig#selectStrategy`
+- opt: 移除过时的`MybatisSqlSessionFactoryBean#typeEnumsPackage`
+- opt: 优化`DdlHelper`资源加载(不再依赖`Spring`或者其他实现)
+- opt: 去除`DdlHelper`中`getScriptRunner`方法指定的字符集编码
+- doc: 修正`DdlHelper`中注释错误
+由于`jsqlParser`5.0版本与5.1版本升级不兼容性不是很大，计划后期移除`mybatis-plus-jsqlparser-5.0`支持模块。
+多版本支持相对来说比较麻烦，后期只维护`mybatis-plus-jsqlparser-4.9` 与 `mybatis-plus-jsqlparser`(保持最新版跟进,直到再提升jdk)
+
 ## [v3.5.10.1] 2025.01.13
 - fix: 修复动态节点处理错误